Create KthLargest.java
Find the Kth largest element in the given list of elements.
The method to find it using Java built-in API is commented out. This could be used to verify the answer.
The traditional way uses merge sort to first sort the list/array before looking for the element.

To find the Kth smallest/largest, play with the ">" and "<" signs in the sorting algorithm.